rename signer to buyer for swap requests
and make the buyer param optional. These will only be needed if routes include is requested. The request does not have to be for the "signer", it can be for any address or contract that could be used for routing request, which is why we are changing the name to "buyer"